+++ gateway/tests/entity_caching/mod.rs
@@ -61,7 +61,7 @@ where
     let federated_schema = {
         let mut subgraphs = graphql_composition::Subgraphs::default();
         subgraphs
-            .ingest_str(subgraph_schema, "the-subgraph", subgraph_url)
+            .ingest_str(subgraph_schema, "the-subgraph", Some(subgraph_url))
             .unwrap();
         let graph = graphql_composition::compose(&subgraphs).into_result().unwrap();
         graphql_composition::render_federated_sdl(&graph).unwrap()